+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- Metric SDK adds `Enricher` API for applying baggage attributes as metric labels in request context. (#XXX) //TODO: add PR number
+
 ## [0.16.0] - 2020-01-13
 
 ### Added